传参数

+++ TeneasyChatSDKUI_Android/src/main/java/com/teneasy/chatuisdk/ui/http/bean/Consults.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 
 data class Consults (
 
-  @SerializedName("consultId" ) var consultId : Int?             = null,
+  @SerializedName("consultId" ) var consultId : Long?             = null,
   @SerializedName("name"      ) var name      : String?          = null,
   @SerializedName("guide"     ) var guide     : String?          = null,
   @SerializedName("Works"     ) var Works     : ArrayList<Works> = arrayListOf(),
